Incluido algumas funcoes para analisar valores constantes, incluido declaracao friend no parser para possibilitar testes de funcoes separadas

+++ FluffyScript/src/fluffy/fl_defs.h
@@ -22,7 +22,7 @@ namespace fluffy {
 	typedef unsigned long long		U64;
 
 	typedef float					Fp32;
-	typedef float					Fp64;
+	typedef double					Fp64;
 
 	typedef void					Void;
 	typedef void* Pointer;
